requirements.txt
# Core Dependencies
crewai==0.28.8
crewai_tools==0.1.6
langchain_community==0.0.29
langchain-core==0.1.53
langchain-google-genai==0.0.11
python-dotenv>=1.0.0
markdown>=3.5
requests>=2.31.0
reportlab>=4.0.0

# RAG Dependencies
pypdf>=4.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# Development Dependencies (optional)
# Install with: pip install pytest pytest-mock
# pytest>=8.0.0
# pytest-mock>=3.12.0
